app/requirements.txt
Flask
gunicorn
tensorflow==2.1.0
google-cloud-storage
numpy
nibabel
pillow
six
sklearn
scikit-image
webcolors
plotly
pandas
matplotlib
h5py>=2.9
fslpy
gast==0.2.2
statsmodels